Non-blocking connection creation
Create connections while not holding the mutex so that multiple connections can be created at once, and so that any slowdown with making connections doesn't stop existing connections from being used. We optimistically treat the connection as created (by incrementing `@created`) so it's still impossible to create more than the max number of connections.

This should ensure that the existing connections remain accessible even while a new connection is being created.